+++ Project/Ken4lowEngine.vcxproj
@@ -120,6 +120,7 @@ copy "$(WindowsSdkDir)bin\$(TargetPlatformVersion)\x64\dxil.dll" "$(TargetDir)dx
   <ItemGroup>
     <ClCompile Include="ApplicationLayer\Character\CharacterWorld\BulletEnemySoABridge.cpp" />
     <ClCompile Include="ApplicationLayer\Scene\DebugScene\DebugActorRegistration.cpp" />
+    <ClCompile Include="Engine\Editor\EditorContentBrowserPanel.cpp" />
     <ClCompile Include="Engine\Physics\Collision\Specialized\BulletEnemyCollisionSoA.cpp" />
     <ClCompile Include="Engine\Scene\Actor\Core\ActorWorld_ComponentEdit.cpp" />
     <ClCompile Include="Engine\Scene\Actor\Core\ActorWorld_Draw.cpp" />
@@ -839,9 +840,14 @@ copy "$(WindowsSdkDir)bin\$(TargetPlatformVersion)\x64\dxil.dll" "$(TargetDir)dx
     <ClInclude Include="ApplicationLayer\Scene\DebugScene\BulletEnemySoABenchmarkPanel.h" />
     <ClInclude Include="ApplicationLayer\Scene\DebugScene\DebugActorRegistration.h" />
     <ClInclude Include="ApplicationLayer\UISystem\UiSpriteFactory.h" />
+    <ClInclude Include="Engine\Editor\ActorWorldEditorBridge.h" />
+    <ClInclude Include="Engine\Editor\EditorAssetRegistryV2.h" />
+    <ClInclude Include="Engine\Editor\EditorContentBrowserPanel.h" />
     <ClInclude Include="Engine\Editor\EditorContext.h" />
+    <ClInclude Include="Engine\Editor\EditorHierarchyPanel.h" />
     <ClInclude Include="Engine\Editor\EditorPanelIds.h" />
     <ClInclude Include="Engine\Editor\EditorShell.h" />
+    <ClInclude Include="Engine\Editor\EditorViewportController.h" />
     <ClInclude Include="Engine\Math\MathUtil.h" />
     <ClInclude Include="Engine\Physics\Collision\Specialized\BulletEnemyCollisionSoA.h" />
     <ClInclude Include="Engine\Scene\Actor\Factory\ActorFactory.h" />
