+++ Project/ken4low_editor_layout_v2.ini
@@ -1,30 +1,30 @@
 [Window][Toolbar]
 Pos=0,26
-Size=1920,53
+Size=1920,62
 Collapsed=0
 DockId=0x00000001,0
 
 [Window][Place Actors]
-Pos=0,81
-Size=230,928
+Pos=0,90
+Size=230,919
 Collapsed=0
 DockId=0x00000003,0
 
 [Window][Main Viewport]
-Pos=232,81
-Size=1289,720
+Pos=232,90
+Size=1289,493
 Collapsed=0
 DockId=0x00000007,0
 
 [Window][Scene]
-Pos=232,81
-Size=1289,720
+Pos=232,90
+Size=1289,493
 Collapsed=0
 DockId=0x00000007,1
 
 [Window][World Outliner]
-Pos=1523,81
-Size=397,388
+Pos=1523,90
+Size=397,379
 Collapsed=0
 DockId=0x00000009,0
 
@@ -35,14 +35,14 @@ Collapsed=0
 DockId=0x0000000A,0
 
 [Window][Content Browser]
-Pos=232,803
-Size=1289,206
+Pos=232,585
+Size=1289,424
 Collapsed=0
 DockId=0x00000008,0
 
 [Window][Output Log]
-Pos=232,803
-Size=1289,206
+Pos=232,585
+Size=1289,424
 Collapsed=0
 DockId=0x00000008,1
 
@@ -101,14 +101,14 @@ Collapsed=0
 DockId=0x00000008
 
 [Window][PhysicsWorld Debug]
-Pos=232,803
-Size=1289,206
+Pos=232,585
+Size=1289,424
 Collapsed=0
 DockId=0x00000008,2
 
 [Window][GPU Particle Editor]
-Pos=232,803
-Size=1289,206
+Pos=232,585
+Size=1289,424
 Collapsed=0
 DockId=0x00000008,3
 
@@ -118,20 +118,19 @@ Size=1920,983
 Collapsed=0
 
 [Window][Debug##Default]
-Pos=12,715
-Size=335,32
+Pos=3,680
+Size=226,43
 Collapsed=0
 
 [Window][FadeManager]
-Pos=9,677
+Pos=10,679
 Size=422,688
 Collapsed=1
 
 [Window][AnimationModel Test]
-Pos=7,759
-Size=335,32
+Pos=23,701
+Size=149,122
 Collapsed=0
-DockId=0x0000000B,0
 
 [Window][Actor World]
 Pos=62,98
@@ -144,16 +143,15 @@ Size=273,104
 Collapsed=0
 
 [Docking][Data]
-DockNode          ID=0x0000000B Pos=7,759 Size=335,32 Selected=0x273AB186
 DockSpace         ID=0x08BD597D Window=0x1BBC0F80 Pos=0,26 Size=1920,983 Split=Y
-  DockNode        ID=0x00000001 Parent=0x08BD597D SizeRef=1280,53 Selected=0x0C01D6D5
-  DockNode        ID=0x00000002 Parent=0x08BD597D SizeRef=1280,665 Split=X
+  DockNode        ID=0x00000001 Parent=0x08BD597D SizeRef=1280,62 Selected=0x0C01D6D5
+  DockNode        ID=0x00000002 Parent=0x08BD597D SizeRef=1280,919 Split=X
     DockNode      ID=0x00000003 Parent=0x00000002 SizeRef=230,665 Selected=0xDDC1CAE2
     DockNode      ID=0x00000004 Parent=0x00000002 SizeRef=1048,665 Split=X
       DockNode    ID=0x00000005 Parent=0x00000004 SizeRef=1289,665 Split=Y
-        DockNode  ID=0x00000007 Parent=0x00000005 SizeRef=774,457 CentralNode=1 Selected=0x6E0FC501
-        DockNode  ID=0x00000008 Parent=0x00000005 SizeRef=774,206 Selected=0x78F77AC4
+        DockNode  ID=0x00000007 Parent=0x00000005 SizeRef=774,493 CentralNode=1 Selected=0x6E0FC501
+        DockNode  ID=0x00000008 Parent=0x00000005 SizeRef=774,424 Selected=0x3DF3100E
       DockNode    ID=0x00000006 Parent=0x00000004 SizeRef=397,665 Split=Y
-        DockNode  ID=0x00000009 Parent=0x00000006 SizeRef=272,278 Selected=0x49BD536E
-        DockNode  ID=0x0000000A Parent=0x00000006 SizeRef=272,385 Selected=0x5DB496A6
+        DockNode  ID=0x00000009 Parent=0x00000006 SizeRef=272,379 Selected=0x49BD536E
+        DockNode  ID=0x0000000A Parent=0x00000006 SizeRef=272,538 Selected=0x5DB496A6
